extrat opcode list
