+++ gb-roms/src/lib.rs
@@ -2,4 +2,4 @@ pub mod header;
 pub mod opcode;
 
 pub use header::{Header, RawHeader};
-pub use opcode::{Opcode, OpcodeGenerator};
+pub use opcode::{list::Opcode, OpcodeGenerator};
